% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_figures.R
\name{axe}
\alias{axe}
\alias{s_axe}
\title{Coordinates of an axe}
\usage{
axe(size.x = 15, pos.x = 0, pos.y = 0, lty = 1, lwd = 3, shd = NA)

s_axe(size.x = 15, pos.x = 0, pos.y = 0, lty = 1, lwd = 3, shd = NA)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse within which the figure is inscribed. Default is 15}

\item{pos.x}{numeric, define the position on the x axis. Default is 0}

\item{pos.y}{numeric, define the position on the y axis. Default is 0}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line)}

\item{lwd}{integer, define the line width of the figure. Default is 3}

\item{shd}{character, define the color of the figure. Default is NA, which results in a transparent figure}
}
\value{
Return the coordinates for drawing an axe

Return the coordinates for drawing a single axe
}
\description{
Define the coordinates for drawing an axe
}
\section{Functions}{
\itemize{
\item \code{s_axe()}: Coordinates of a single axe

Define the coordinates for drawing a single axe, to be used in shape()

}}
\examples{
# return the default coordinates for drawing an axe
axe()
# change the coordinates for drawing a smaller single axe
axe(size.x = 5)
# return the default coordinates for drawing single axe
s_axe()
# change the coordinates for drawing a smaller single axe
s_axe(size.x = 5)
}
