% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composed_figures.R, R/other-figures.R
\name{bow.tie}
\alias{bow.tie}
\title{Define the coordinates of a vertical bow tie}
\usage{
bow.tie(size.x = 10, pos.x = 0, shd = NA, lwd = 3, lty = 1)

bow.tie(size.x = 10, pos.x = 0, shd = NA, lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{}

\item{lty}{}
}
\description{
Define the coordinates of a vertical bow tie

Define the cooordinates for a vertical bow tie
}
