% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.R
\name{change_color.figure}
\alias{change_color.figure}
\title{Change color}
\usage{
\method{change_color}{figure}(obj, ...)
}
\arguments{
\item{obj}{The figure}

\item{...}{other arguments}
}
\description{
Change color
}
