% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.R
\name{change_color}
\alias{change_color}
\title{Method for changing color}
\usage{
change_color(obj, ...)
}
\arguments{
\item{obj}{The figure}

\item{...}{other arguments}
}
\value{
l
}
\description{
Method for changing color
}
