% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{circle}
\alias{circle}
\title{Define the coordinates of a circle}
\usage{
circle(
  size.x = 10,
  size.y = 10,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  rot = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{Length of the x axis}

\item{size.y}{Length of the y axis}

\item{pos.x}{Position on the x axis}

\item{pos.y}{Position on the y axis}

\item{shd}{Color of the object. Deafsult is NA which results in a transparent object}

\item{rot}{Rotation of the ellipse in which the figure is inscribed}

\item{vis}{Integer, indicates whether the object should be visible (1) or not (0). Deafult is visible}

\item{lty}{Border line. Default is 1 (solid), can be dotted (2) or dashed (3)}

\item{lwd}{Width of the border line. Deafult is 3.}
}
\value{
Based on the chosen shape, return an object with the information for plotting the desired design
}
\description{
Define the coordinates of a circle
}
\examples{
Ci sarà
}
