% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{cof.character}
\alias{cof.character}
\title{Cof character}
\usage{
\method{cof}{character}(...)
}
\arguments{
\item{...}{Arguments of the function}
}
\description{
Per generalizzare la funzione per concatenzione di matrici ho definito metodo cof anche per i caratteri
}
