% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{cof.figure}
\alias{cof.figure}
\title{Cof Figure}
\usage{
\method{cof}{figure}(..., name = NULL, single = FALSE)
}
\arguments{
\item{...}{The figures to be concatenated}

\item{name}{character, the name to assign to the newly created figure. Default is NULL (no name).}

\item{single}{logical, define whether the newly created figure is considered as single element to e used in shape() or not. Default is FALSE.}
}
\description{
Cof Figure
}
