% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{com}
\alias{com}
\title{Concatenation of matrices}
\usage{
com(...)
}
\arguments{
\item{...}{the matrices to be concatenated}
}
\value{
Return a matrix composed of 2+ matrices
}
\description{
Hierarchically concatenates 2+ matrices on top of one another. The first matrix is placed on the bottom, the last matrix is placed on top of all other matrices.
}
