% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{concatenation.character}
\alias{concatenation.character}
\title{Concatenation of stuff}
\usage{
\method{concatenation}{character}(...)
}
\arguments{
\item{...}{The matrices to be concatenated}
}
\description{
Concatenation of stuff
}
