% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{correct.matriKS}
\alias{correct.matriKS}
\title{Extract the cell of the correct response}
\usage{
\method{correct}{matriKS}(obj)
}
\arguments{
\item{obj}{}
}
\value{
The cell with the correct response, either Sq9 (9-cell matrix) or Sq5 (4-cell matrix)
}
\description{
Extract the cell of the correct response
}
