% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{correct}
\alias{correct}
\alias{correct.matriks}
\title{Correct response (Method)}
\usage{
correct(obj)

\method{correct}{matriks}(obj)
}
\arguments{
\item{obj}{The matrix}
}
\value{
The correct response of a matriks

The correct response of a matriks
}
\description{
Isolate the correct response from a matriks
}
\section{Methods (by class)}{
\itemize{
\item \code{correct(matriks)}: Correct response

Isolate the correct response from a matriks

}}
\examples{
# apply the size rule on a triangle for creating a matriks with 9 cell
my_mat <- mat_apply(triangle(), mat.type = 9, hrule = "size")
# draw the matriks without the correct response
draw(my_mat, hide = TRUE)
# add the correct response
draw(correct(my_mat))
# apply the reflect rule on a pacman for creating a matriks with 4 cell
my_mat <- mat_apply(pacman(), mat.type = 4, vrule = "reflect")
# draw the matriks without the correct response
draw(my_mat, hide = TRUE)
# add the correct response
draw(correct(my_mat))
# apply the size rule on a triangle for creating a matriks with 9 cell
my_mat <- mat_apply(triangle(), mat.type = 9, hrule = "size")
# draw the matriks without the correct response
draw(my_mat, hide = TRUE)
# add the correct response
draw(correct(my_mat))

# apply the reflect rule on a pacman for creating a matriks with 4 cell
my_mat <- mat_apply(pacman(), mat.type = 4,
                  vrule = "reflect")
# draw the matriks without the correct response
draw(my_mat, hide = TRUE)
# add the correct response
draw(correct(my_mat))
}
