% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decof.R
\name{decof.figure}
\alias{decof.figure}
\title{Split the elements of a figure}
\usage{
\method{decof}{figure}(obj)
}
\arguments{
\item{The}{figure to be split in its single components}
}
\value{
A named list of figures
}
\description{
Return the elements composing a figure
}
\examples{
la salma è la virtù dei morti
}
