% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decof.R
\name{decof}
\alias{decof}
\title{Split the elements of a figure}
\usage{
decof(obj)
}
\arguments{
\item{The}{figure to be split in its single components}
}
\value{
A named list of figures
}
\description{
Return the elements composing a figure
}
\examples{
\dontrun{
# apply the size rule on a triangle for creating a matriks with 9 cell
my_mat = mat_apply(triangle(), mat.type = 9,
hrule = "size")
# Return the figures composing the first cell of the matriks
decof(my_mat)
}
}
