% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{diagline}
\alias{diagline}
\title{Define the coordinates of the main diagonal line}
\usage{
diagline(
  size.x = list(sqrt(square()$size.x[[1]]^2/2)),
  size.y = list(sqrt(square()$size.x[[1]]^2/2)),
  pos.x = 0,
  pos.y = 0,
  lty = 1,
  lwd = 3,
  rotation = pi - pi/4,
  vis = 1
)
}
\arguments{
\item{vis}{}
}
\description{
Define the coordinates of the main diagonal line
}
