% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{dice}
\alias{dice}
\title{Define the coordinates of a dice with 4 dots}
\usage{
dice(pos.x = 13, pos.y = 13, shd = "black", lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a dice with 4 dots
}
