% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{difference.matriks}
\alias{difference.matriks}
\title{Difference distractors}
\usage{
\method{difference}{matriks}(obj, seed = 666, ...)
}
\arguments{
\item{obj}{the object}

\item{seed}{the seed}

\item{...}{Other arguments}
}
\description{
Difference distractors
}
