% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{difference}
\alias{difference}
\title{Difference (method)}
\usage{
difference(obj, seed, ...)
}
\arguments{
\item{obj}{the matrikes}

\item{seed}{seed}

\item{...}{other arguments}
}
\value{
An object
}
\description{
Difference distractors
}
