% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{down.petal}
\alias{down.petal}
\title{Define the coordinates of a down petal}
\usage{
down.petal(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a down petal
}
\description{
Define the coordinates of the circle sections composing a down petal
}
\examples{
\dontrun{
# return the default coordinates for drawing a down petal
down.petal()

# change the line type of the down petal

down.petal(lty = 3)
}
}
