% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.responses}
\alias{draw.responses}
\title{Draw responses}
\usage{
\method{draw}{responses}(
  obj,
  main = NULL,
  canvas = TRUE,
  hide = FALSE,
  bg = "white",
  mar = c(1, 1, 1, 1),
  xlim = 16,
  distractors = NULL,
  print = FALSE,
  ...
)

\method{draw}{responses}(
  obj,
  main = NULL,
  canvas = TRUE,
  hide = FALSE,
  bg = "white",
  mar = c(1, 1, 1, 1),
  xlim = 16,
  distractors = NULL,
  print = FALSE,
  ...
)
}
\arguments{
\item{obj}{The figure to be drawn. Can be a single figure, a matrix, or the responses}

\item{main}{logical, print the title of the drawing. Default is FALSE}

\item{canvas}{logical, draw the figure on a new canvas. Default is TRUE}

\item{hide}{logical, hide the cell corresponding to the correct response. Default is FALSE}

\item{bg}{character, define the color background. Default is white}

\item{mar}{numeric vector, change margins of the canvas}

\item{xlim}{numeric, change the length of the x axis}

\item{distractors}{the distractors}

\item{print}{the prrint}

\item{...}{Other arguments}

\item{n.figure}{integer, define the number of cells of the matrix. Default is 9}
}
\value{
a list

a list
}
\description{
Draw the response list

Draw the response list
}
