% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{ellipse}
\alias{ellipse}
\title{Coordinates of an ellipse}
\usage{
ellipse(
  size.x = 10,
  size.y = 7,
  rot = 0,
  shd = NA,
  pos.x = 0,
  pos.y = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse within which the figure is inscribed. Default is 10}

\item{size.y}{numeric, define the semi-minor axis of the ellipse within which the figure is inscribed. Default is 7}

\item{rot}{define the rotation. Default is 0}

\item{shd}{character, define the shading of the figure. Default is NA which results in a transparent figure}

\item{pos.x}{numeric, position on the x axis. Default is 0}

\item{pos.y}{numeric, position the y axis, Default is 0}

\item{vis}{Visibility of the figure. Default is 1, making the figure visible. To hide the figure, change it to 0}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3}
}
\value{
Return the coordinates for drawing a ellipse
}
\description{
Define the coordinates for drawing an ellipse
}
\examples{
# return the default coordinates for drawing an ellipse
ellipse()
# change the coordinates for drawing a smaller ellipse
ellipse(size.x = 5, size.y = 3)
}
