% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{fill}
\alias{fill}
\title{Change the shading of a figure}
\usage{
fill(fig, n, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the color of the shading. Default is 1 (white). Other options are 2 (grey) and 3 (black)}

\item{...}{}
}
\description{
Change the shading of a figure
}
