% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_arc.R
\name{h.arc.left.up}
\alias{h.arc.left.up}
\title{Define the coordinates of the horizontal left up arch}
\usage{
h.arc.left.up(
  size.x = square()$size.x[[1]]/2,
  size.y = square()$size.y[[1]]/2,
  lty = 1,
  lwd = 3,
  vis = 1,
  pos.x = 0,
  pos.y = 0
)
}
\value{
Return the horizontal arc left up object
}
\description{
Define the coordinates of the horizontal left up arch
}
\examples{
h.arc.left.up()
}
