% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_arc.R
\name{h_arc_right_up}
\alias{h_arc_right_up}
\title{Coordinates of a horizontal right up arch}
\usage{
h_arc_right_up(
  size.x = square()$size.x[[1]]/2,
  size.y = size.x,
  lty = 1,
  lwd = 3,
  vis = 1,
  pos.x = 0,
  pos.y = 0
)
}
\arguments{
\item{size.x}{The length of the x-axis. Default is square()$size.x[[1]]/2.}

\item{size.y}{The length of the x-axis. Default is size.x.}

\item{lty}{The lime type, default is 1 (solid line).}

\item{lwd}{The line width. Default is 3}

\item{vis}{Visibility of the figure. Deafult is 1, making the figure visible. To hide the figure, change it to 0.}

\item{pos.x}{Position on the x axis. Default is 0.}

\item{pos.y}{Position on the y axis, Default is 0.}
}
\value{
The coordinates of the right up arch
}
\description{
Define the coordinates of the right up arch of a circle
}
\examples{
\dontrun{
# default coordinates of the right up arch

h_arc_right_up()

}
}
