% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{hexagon}
\alias{hexagon}
\title{Define the coordinates of an hexagon}
\usage{
hexagon(
  size.x = 15,
  size.y = 15,
  rot = 0,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{lwd}{}
}
\description{
Define the coordinates of an hexagon
}
