% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_cell.R
\name{hide}
\alias{hide}
\title{Functions to modify the visible objects in a cell}
\usage{
hide(obj, index)
}
\arguments{
\item{obj}{The cell of a matrix}

\item{index}{The index of the element to hide/show/replace}

\item{replacement}{The object with which an element should be replaced}
}
\description{
Functions to modify the visible objects in a cell
}
\examples{
Arrivano
}
