% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hline}
\alias{hline}
\title{Define the coordinates of an horizontal line}
\usage{
hline(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = sqrt(square()$size.y[[1]]^2/2),
  pos.x = 0,
  pos.y = 0,
  lty = 1,
  lwd = 3,
  vis = 1
)
}
\arguments{
\item{vis}{}
}
\description{
Define the coordinates of an horizontal line
}
