% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{horizontal_s}
\alias{horizontal_s}
\title{Coordinates of an horizontal S}
\usage{
horizontal_s(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
Return the coordinates for drawing an horizontal S
}
\description{
Define the coordinates of an horizontal S
}
\examples{
\dontrun{
# default coordinates of the horizontal S
horizontal_s()
# change the line type
horizontal_s(lty = 2)
}
}
