% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic.matriks}
\alias{ic.matriks}
\title{Incomplete correlate distractors}
\usage{
\method{ic}{matriks}(obj, ...)
}
\arguments{
\item{obj}{object of class obj}

\item{...}{Other arguments}
}
\value{
A list of length three
}
\description{
Generates the incomplete correlate distractors
}
