% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R, R/ic_flip.R, R/ic_inc.R, R/ic_neg.R,
%   R/ic_size.R
\name{ic}
\alias{ic}
\alias{ic.matriks}
\alias{ic_flip}
\alias{ic_flip.matriks}
\alias{ic_inc}
\alias{ic_inc.matriks}
\alias{ic_neg}
\alias{ic_neg.matriks}
\alias{ic_size}
\alias{ic_size.matriks}
\title{Incomplete correlate distractors (method)}
\usage{
ic(obj)

\method{ic}{matriks}(obj, ...)

ic_flip(obj, ...)

\method{ic_flip}{matriks}(obj, ...)

ic_inc(obj, ...)

\method{ic_inc}{matriks}(obj, ...)

ic_neg(obj, ...)

\method{ic_neg}{matriks}(obj, ...)

ic_size(obj, ...)

\method{ic_size}{matriks}(obj, ...)
}
\arguments{
\item{obj}{matriks, The matriks for which the distractor is generated}

\item{...}{other arguments}
}
\value{
An object of class responses of length 4, which contains the incomplete correlate distractors of a matriks (IC-Inc, IC-Flip, IC-Neg, IC-Size). If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class responses of length 4, which contains the incomplete correlate distractors of a matriks. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate flip distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate flip distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate incomplete distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate incomplete distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate negative distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate negative distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate size distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class figure that is the incomplete correlate size distractor of a matrix. If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.
}
\description{
Generate incomplete correlate flip distractor from a matriks
}
\section{Methods (by class)}{
\itemize{
\item \code{ic(matriks)}: Incomplete correlate distractors

Generate incomplete correlate flip distractor from a matriks

}}
\section{Functions}{
\itemize{
\item \code{ic_flip()}: Incomplete correlate flip distractor (method)

Generate incomplete correlate flip distractor from a matriks

\item \code{ic_flip(matriks)}: Incomplete correlate flip distractor

Generate incomplete correlate flip distractor from a matriks

\item \code{ic_inc()}: Incomplete correlate incomplete distractor (method)

Generate incomplete correlate incomplete distractor from a matriks

\item \code{ic_inc(matriks)}: Incomplete correlate incomplete distractor

Generate incomplete correlate incomplete distractor from a matriks

\item \code{ic_neg()}: Incomplete correlate negative distractor (method)

Generate incomplete negative incomplete distractor from a matriks

\item \code{ic_neg(matriks)}: Incomplete correlate negative distractor

Generate incomplete negative incomplete distractor from a matriks

\item \code{ic_size()}: Incomplete correlate size distractor (method)

Generate incomplete size incomplete distractor from a matriks

\item \code{ic_size(matriks)}: Incomplete correlate size

Generate incomplete correlate size distractor of a matrix

}}
\examples{
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
m2 <- mat_apply(dot(), "shade")
mat <- com(m1, m2)
# draw the matrix
draw(mat)
# draw the incomplete correlate distractors
draw(ic(mat))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
m2 <- mat_apply(dot(), "shade")
mat <- com(m1, m2)
# draw the matrix
draw(mat)
# draw the incomplete correlate distractors
draw(ic(mat))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
# draw the matrix
draw(m1)
# draw the incomplete correalate flip distractor
draw(ic_flip(m1))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
# draw the matrix
draw(m1)
# draw the incomplete correalate flip distractor
draw(ic_flip(m1))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
m2 <- mat_apply(dot(), "shade")
mat <- com(m1, m2)
# draw the matrix
draw(mat)
# draw the incomplete correlate incomplete distractor
draw(ic_inc(mat))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
m2 <- mat_apply(dot(), "shade")
mat <- com(m1, m2)
# draw the matrix
draw(mat)
# draw the incomplete correlate incomplete distractor
draw(ic_inc(mat))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
# draw the matrix
draw(m1)
# draw the incomplete correlate negative distractor
draw(ic_neg(m1))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
# draw the matrix
draw(m1)
# draw the incomplete correlate negative distractor
draw(ic_neg(m1))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
# draw the incomplete correlate size distractor
draw(ic_size(m1))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
# draw the incomplete correlate size distractor
draw(ic_size(m1))
}
