% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_flip.R
\name{ic_flip.matriks}
\alias{ic_flip.matriks}
\title{flipative incomplete correlate}
\usage{
\method{ic_flip}{matriks}(obj, ...)
}
\arguments{
\item{obj}{The matriks}

\item{...}{Other arguments}
}
\value{
A list
}
\description{
Change color to the correct response for creating the ic flip distractors
}
