% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_inc.R
\name{ic_inc.matriks}
\alias{ic_inc.matriks}
\title{incative incomplete correlate}
\usage{
\method{ic_inc}{matriks}(obj, ...)
}
\arguments{
\item{obj}{The matriks}

\item{...}{Other arguments}
}
\value{
A list
}
\description{
Change color to the correct response for creating the ic inc distractors
}
