% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_neg.R
\name{ic_neg.matriks}
\alias{ic_neg.matriks}
\title{Negative incomplete correlate}
\usage{
\method{ic_neg}{matriks}(obj, ...)
}
\arguments{
\item{obj}{The matriks}

\item{...}{Other arguments}
}
\value{
A list
}
\description{
Change color to the correct response for creating the ic neg distractors
}
