% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_size.R
\name{ic_size.matriks}
\alias{ic_size.matriks}
\title{sizeative incomplete correlate}
\usage{
\method{ic_size}{matriks}(obj, ...)
}
\arguments{
\item{obj}{The matriks}

\item{...}{Other arguments}
}
\value{
A list
}
\description{
Change color to the correct response for creating the ic size distractors
}
