% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{identity.figure}
\alias{identity.figure}
\title{Identity figure}
\usage{
\method{identity}{figure}(fig, ...)
}
\arguments{
\item{fig}{Vector of figures obtained with the concatenation of figures function (`cof()`). Three figures are needed.}

\item{...}{Other arguments}
}
\description{
Identity figure
}
