% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{identity}
\alias{identity}
\alias{identity.figure}
\title{Identity rule (Method)}
\usage{
identity(fig, ...)

\method{identity}{figure}(fig, ...)
}
\arguments{
\item{fig}{Vector of figures obtained with the concatenation of figures function (`cof()`). Three figures are needed.}

\item{...}{Other arguments}
}
\value{
An object composed of figures combined according to an identity rule
}
\description{
Apply an identity rule to the figures in a matrix (i.e., no changes)
}
\section{Methods (by class)}{
\itemize{
\item \code{identity(figure)}: Identity figure

}}
\examples{
# generate a matrix with 9 squares
draw(mat_apply(square(), hrules = "identity"))
# generate a matrix with 9 squares
draw(mat_apply(square(), hrules = "identity"))
}
