% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{left.petal}
\alias{left.petal}
\title{Define the coordinates of a the left petal}
\usage{
left.petal(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a the left petal
}
