% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{logic.figure}
\alias{logic.figure}
\title{Logic figure}
\usage{
\method{logic}{figure}(fig, n = 1, rule = "logic", seed = 1, ...)
}
\arguments{
\item{fig}{Vector of figures obtained with the concatenation of figures function (`cof()`). Three figures are needed.}

\item{n}{???}

\item{rule}{Define the logic rule to be applied, either `AND`, `OR`, `XOR`}

\item{seed}{Set the random seed so that the permutations are consistent}

\item{...}{Other arguments}
}
\description{
Logic figure
}
