% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{logic}
\alias{logic}
\alias{logic.figure}
\title{Logical rules (Method)}
\usage{
logic(fig, n, rule, seed, ...)

\method{logic}{figure}(fig, n = 1, rule = "logic", seed = 1, ...)
}
\arguments{
\item{fig}{Vector of figures obtained with the concatenation of figures function (`cof()`). Three figures are needed.}

\item{n}{integer, defines the elements of the logical expression. n=1 and n=2 are the concatenations of figures to which the logical operation is applied. n=3 is the result of the operation.}

\item{rule}{character, logic rule to be applied, either `AND`, `OR`, `XOR`}

\item{seed}{integer, Set the random seed so that the permutations are consistent}

\item{...}{Other arguments}
}
\value{
An object that is the logical combination of the figures

An object that is the logical combination of the figures
}
\description{
Apply logical rules (intersection--AND, union--OR, symmetrical difference--XOR) to a concatenation of figures
}
\section{Methods (by class)}{
\itemize{
\item \code{logic(figure)}: Logical rules

Apply logical rules (intersection--AND, union--OR, symmetrical difference--XOR) to a concatenation of figures

}}
\examples{
draw(logic(cof(square(), malta(), circle()), "AND"))

draw(logic(cof(square(), malta(), circle()), "AND"))

}
