% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{luck.4}
\alias{luck.4}
\title{Coordinates of a luck composed of 4 lines}
\usage{
luck.4(size.x = 10, size.y = 7, lwd = 3, lty = 1)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 10}

\item{size.y}{numeric, define the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 7}

\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing an X
}
\description{
Define the coordinates for drawing of a luck composed of 4 lines
}
\examples{
\dontrun{
# default coordinates of an luck composed of 4 lines

luck.4()

# draw a luck composed of 4 lines with different lty

draw(luck.4(lty = 2))

}
}
