% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-figures.R
\name{malta}
\alias{malta}
\title{Define the cooordinates for a Malta cross}
\usage{
malta(size.x = 10, pos.x = 0, shd = NA, lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the cooordinates for a Malta cross
}
