% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{margin.figure}
\alias{margin.figure}
\title{Margin}
\usage{
\method{margin}{figure}(obj, n, rule, ...)
}
\arguments{
\item{obj}{the figure}

\item{n}{ineteger, either the linetype or the linewidth}

\item{rule}{character, lty changes the linetype (1 = solid, 2 = dashed, 3 = dotted), lwd changes the linewdith}

\item{...}{Other arguments}
}
\value{
A figure with changed margins
}
\description{
Chnage the margin of the figure
}
\examples{
\dontrun{

# draw default triangle
draw(triangle())

# change the linetype
draw(margin(triangle(), 2))

}
}
