% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_apply.R
\name{mat_apply.matriks}
\alias{mat_apply.matriks}
\title{Split the elements of a figure}
\usage{
mat_apply.matriks(Sq1, mat.type = 9, hrules = "identity", vrules = "identity")
}
\arguments{
\item{obj}{The figure to be split in its single components}
}
\value{
A named list of figures
}
\description{
Return the elements composing a figure
}
