% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_apply.R
\name{mat_apply}
\alias{mat_apply}
\title{Split the elements of a figure}
\usage{
mat_apply(Sq1, mat.type = 9, hrules = "identity", vrules = "identity")
}
\arguments{
\item{obj}{The figure to be split in its single components}
}
\value{
A named list of figures
}
\description{
Return the elements composing a figure
}
\examples{
\dontrun{
# apply the size rule on a triangle for creating a matriks with 9 cell
my_mat = mat_apply(triangle(), mat.type = 9,
hrule = "size")
}
}
