% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{miley}
\alias{miley}
\alias{s_miley}
\title{Define the coordinates of a miley}
\usage{
miley(lwd = 3, lty = 1)

s_miley(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line)}
}
\value{
Return the coordinates for drawing the petals composing a miley

Return the coordinates for drawing the petals composing a single miley
}
\description{
Define the coordinates for drawing the petals composing a miley
}
\section{Functions}{
\itemize{
\item \code{s_miley()}: Define the coordinates a single miley

Define the coordinates for drawing the petals composing a single miley, to be used in shape()

}}
\examples{
# return the default coordinates for drawing a right petal
miley()
# change the line type of the right petal
miley(lty = 3)
# return the default coordinates for drawing the petals composing a single miley
s_miley()
# change the line type of the single miley
s_miley(lty = 3)
}
