% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_figures.R
\name{ninja}
\alias{ninja}
\alias{s_ninja}
\title{Coordinates of a ninja star}
\usage{
ninja(size.x = 10, size.y = 15, shd = "black", lwd = 3, lty = 0)

s_ninja(size.x = 10, size.y = 15, shd = "black", lwd = 3, lty = 0)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse within which the figure is inscribed. Default is 10}

\item{size.y}{numeric, define the semi-minor axis of the ellipse within which the figure is inscribed. Default is 15}

\item{shd}{character, define the shading of the figure. Default is black}

\item{lwd}{integer, define the line width of the figure. Default is 3}

\item{lty}{integer, define the line type of the figure, default is 0}
}
\value{
Return the coordinates for drawing a ninja star

Return the coordinates for drawing a single ninja
}
\description{
Define the coordinates for drawing a ninja star (composed of two lucks)
}
\section{Functions}{
\itemize{
\item \code{s_ninja()}: Coordinates of a single ninja

Define the coordinates for drawing a single ninja star (composed of two lucks), to be used in shape()

}}
\examples{
# return the default coordinates for drawing a ninja
ninja()
# change the shade of the ninja
ninja(shd = "grey", lty = 0)
# return the default coordinates for drawing a single ninja
s_ninja()
# change the shade of the single ninja
s_ninja(shd = "grey", lty = 0)
}
