% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections.R
\name{pizza.2.rev}
\alias{pizza.2.rev}
\title{Define the coordinates of an inverse pizza with two slices}
\usage{
pizza.2.rev(
  size.x = 15,
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{lwd}{}
}
\description{
Define the coordinates of an inverse pizza with two slices
}
