% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections.R
\name{pizza.4}
\alias{pizza.4}
\title{Define the coordinates of a pizza with four slices}
\usage{
pizza.4(size.x = 15, shd = NA, lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a pizza with four slices
}
