% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{rectangle}
\alias{rectangle}
\title{Coordinates of a rectangle}
\usage{
rectangle(
  size.x = 10,
  size.y = 8,
  rot = pi/4,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  lwd = 3,
  lty = 1,
  vis = 1
)
}
\arguments{
\item{size.x}{An integer giving the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 10.}

\item{size.y}{An integer or a vector giving the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 8.}

\item{rot}{Rotation of the figure. Default is pi/4.}

\item{pos.x}{Position on the x axis. Default is 0.}

\item{pos.y}{Position on the y axis, Default is 0.}

\item{shd}{Color of the figure Default is NA which results in a transparent figure}

\item{lwd}{An integer defining the width of the border line. Default is 3.}

\item{lty}{An integer defining the border line. Default is 1 (solid), can be dotted (2) or dashed (3)}

\item{vis}{Integer, indicates whether the object should be visible (1) or not (0). Default is 1}
}
\value{
Return the coordinates for drawing a rectangle
}
\description{
Define the coordinates  of the ellipse within which a rectangle can be inscribed.
}
\examples{
\dontrun{
# return the default coordinates for drawing a rectangle
rectangle()

# change the coordinates for drawing a smaller rectangle

rectangle(size.x = 5, size.y = 10)
}
}
