% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{reflect.figure}
\alias{reflect.figure}
\title{Reflect figure}
\usage{
\method{reflect}{figure}(fig, n = 2, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the reflection. Default is 2 which corresponds to a 180 degree (pi) reflection}

\item{...}{Other arguments}
}
\description{
Reflect figure
}
