% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{reflect}
\alias{reflect}
\alias{reflect.figure}
\title{Reflection rule (Method)}
\usage{
reflect(fig, n, ...)

\method{reflect}{figure}(fig, n = 2, ...)
}
\arguments{
\item{fig}{The figure to be reflected}

\item{n}{integer, defines the angle of the rotation. Default is 2}

\item{...}{Other arguments}
}
\value{
A figure of class figure with different rotation coordinates

A figure of class figure with different rotation coordinates
}
\description{
Apply a rotation of \eqn{\pi} to a figure.
}
\section{Methods (by class)}{
\itemize{
\item \code{reflect(figure)}: Reflect a figure

Apply a rotation of \eqn{\pi} to a figure.

}}
\examples{
# default pacman
draw(pacman())

# apply the default reflection on the default pacman
draw(reflect(pacman()))
# default pacman
draw(pacman())

# apply the default reflection on the default pacman
draw(reflect(pacman()))
}
