% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repetition.R
\name{repetition.matriks}
\alias{repetition.matriks}
\title{Repetition distractors}
\usage{
\method{repetition}{matriks}(obj, ...)
}
\arguments{
\item{obj}{object of class obj}

\item{...}{Other arguments}
}
\value{
A list of length three
}
\description{
Generates the repetition distractors
}
