% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repetition.R
\name{repetition}
\alias{repetition}
\alias{repetition.matriks}
\title{Repetition distractors (Method)}
\usage{
repetition(obj, ...)

\method{repetition}{matriks}(obj, ...)
}
\arguments{
\item{obj}{matriks, The matriks for which the distractor is generated}

\item{...}{other arguments}
}
\value{
An object of class responses of length 3, which contains the repetition distractors of a matriks (R-Left, R-Top, R-Diag). If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class responses of length 3, which contains the repetition distractors of a matriks (R-Left, R-Top, R-Diag). If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.
}
\description{
Generate repetition distractors from a matriks
}
\section{Methods (by class)}{
\itemize{
\item \code{repetition(matriks)}: Repetition distractors (Method)

Generate repetition distractors from a matriks

}}
\examples{
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
m2 <- mat_apply(dot(), "shade")
mat <- com(m1, m2)
# draw the matrix
draw(mat)
# draw the repetition distractors
draw(repetition(mat))
# create a matrix
m1 <- mat_apply(pacman(), hrules = "lty")
m2 <- mat_apply(dot(), "shade")
mat <- com(m1, m2)
# draw the matrix
draw(mat)
# draw the repetition distractors
draw(repetition(mat))
}
