% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{replace.figure}
\alias{replace.figure}
\title{Replace objects}
\usage{
\method{replace}{figure}(obj, index, replacement, visible = FALSE)
}
\arguments{
\item{obj}{The figure of a matrix}

\item{index}{The index of the element to replace}

\item{replacement}{The name of the figure with which the original one is replaced}

\item{visible}{If true it will replace only the visible figure}
}
\value{
An object with a changed figure

Return an object with a different figure
}
\description{
Replace an object in a figure with another object
}
\examples{
\dontrun{
# concanate three figures into an object
my_shapes = cof(square(), triangle(), slice())
# draw object
draw(my_shapes)
# replace the square with a gray pacman
draw(replace(my_shapes, 1, pacman(shd = "grey")))
}
}
