% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_list.R
\name{response_list.matriks}
\alias{response_list.matriks}
\title{Response list}
\usage{
\method{response_list}{matriks}(obj, seed = 666, ...)
}
\arguments{
\item{obj}{the matrix}

\item{seed}{the seed}

\item{...}{other arguments}
}
\value{
A list
}
\description{
Response list
}
