% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rotate.figure}
\alias{rotate.figure}
\title{Rotate figure}
\usage{
\method{rotate}{figure}(fig, n = 4, rule = "rotation", ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the angle of the rotation. Default is 4, which corresponds to a rotation of pi/4.}

\item{rule}{Define the rotation rule. Default is counterclockwise.`rule = "inv"` forces a clockwise rotation}

\item{...}{Other arguments}
}
\description{
Rotate figure
}
