% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rotate}
\alias{rotate}
\title{Rotates a figure}
\usage{
rotate(fig, n, rule, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the angle of the rotation. Default is 4, which corresponds to a rotation of $\frac{\pi}{4}$.}

\item{rule}{Define the rotation rule. Default is anticlockwise.`rule = "inv"` forces a clockwise rotation}

\item{...}{Other arguments}
}
\description{
Rotates a figure
}
\examples{
one day not todaty
}
