% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-figures.R
\name{s.axe}
\alias{s.axe}
\title{Define the cooordinates for a single axe (to be used in diff_shapes)}
\usage{
s.axe(size.x = 15, pos.x = 0, pos.y = 0, lty = 1, lwd = 3, shd = NA)
}
\arguments{
\item{shd}{}
}
\description{
Define the cooordinates for a single axe (to be used in diff_shapes)
}
