% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{s.biscuit}
\alias{s.biscuit}
\title{Define the coordinates of a biscuit (to be used in diff_shapes)}
\usage{
s.biscuit(
  pos.x = 0,
  pos.y = 0,
  size.x = 10,
  size.y = 10,
  shd = "black",
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{lwd}{}
}
\description{
Define the coordinates of a biscuit (to be used in diff_shapes)
}
