% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composed_figures.R, R/other-figures.R
\name{s.bow.tie}
\alias{s.bow.tie}
\title{Define the coordinates of a single vertical bow tie (to be used in diff_shapes)}
\usage{
s.bow.tie(size.x = 10, pos.x = 0, shd = NA, lwd = 3, lty = 1)

s.bow.tie(size.x = 10, pos.x = 0, shd = NA, lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{}

\item{lty}{}
}
\description{
Define the coordinates of a single vertical bow tie (to be used in diff_shapes)

Define the cooordinates for a single bow tie (to be used in diff_shapes)
}
