% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{s.flower}
\alias{s.flower}
\title{Define the coordinates a single flower}
\usage{
s.flower(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{The line width. Default is 3.}

\item{lty}{The lime type, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a single flower
}
\description{
Define the coordinates of the circle sections composing a flower, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# return the default coordinates for drawing a single flower
s.flower()

# change the line type of the single flower

s.flower(lty = 3)
}
}
