% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.horizontal.inv}
\alias{s.horizontal.inv}
\title{Define the cooordinates for an horizontal invertred s}
\usage{
s.horizontal.inv(lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for an horizontal invertred s
}
