% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.horizontal}
\alias{s.horizontal}
\title{Define the cooordinates for an horizontal s}
\usage{
s.horizontal(lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for an horizontal s
}
