% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.horizontal.s}
\alias{s.horizontal.s}
\title{Coordinates of a single horizontal S}
\usage{
s.horizontal.s(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
Return the coordinates for drawing a single horizontal S
}
\description{
Define the coordinates for drawing a single horizontal S composed of two arches, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# default coordinates of the single horizontal S
s.horizontal.s()
# change the line type
s.horizontal.s(lty = 2)
}
}
