% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-figures.R
\name{s.malta}
\alias{s.malta}
\title{Define the cooordinates for a single Malta cross (to be used in diff_shapes)}
\usage{
s.malta(size.x = 10, pos.x = 0, shd = NA, lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the cooordinates for a single Malta cross (to be used in diff_shapes)
}
