% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-figures.R
\name{s.maxi}
\alias{s.maxi}
\title{Define the cooordinates for a single maxi (to be used in diff_shapes)}
\usage{
s.maxi(size.x = 8, size.y = 4, pos.x = 0, shd = NA, lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for a single maxi (to be used in diff_shapes)
}
