% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections.R
\name{s.pizza.2}
\alias{s.pizza.2}
\title{Define the coordinates of a single pizza with two slices (to be used in diff_shapes)}
\usage{
s.pizza.2(
  size.x = 15,
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{lwd}{}
}
\description{
Define the coordinates of a single pizza with two slices (to be used in diff_shapes)
}
