% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections_figures.R
\name{s.pizza.2.rev}
\alias{s.pizza.2.rev}
\title{Coordinates of a single inverse pizza with two slices}
\usage{
s.pizza.2.rev(
  size.x = 15,
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is sqrt(square()$ size.x[[1]]^2 /2)}

\item{size.y}{numeric, define the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is size.x.}

\item{pos.x}{numeric, position on the x axis. Default is 0.}

\item{pos.y}{numeric, position the y axis, Default is 0.}

\item{shd}{character, define the shading of the figure. Default is NA which results in a transparent figure}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
The coordinates of two circle sections composing a single pizza with two slices
}
\description{
Define the coordinates of the circle section composing an inverse pizza with two slices, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# default coordinates of the single inverse pizza with two slices
s.pizza.2()
}
}
