% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.s.horizontal.inv}
\alias{s.s.horizontal.inv}
\title{Define the cooordinates for a single inverted horizontal s (to be used in diff shapes)}
\usage{
s.s.horizontal.inv(lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for a single inverted horizontal s (to be used in diff shapes)
}
