% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.s.vertical}
\alias{s.s.vertical}
\title{Define the cooordinates for a single vertical s (to be used in diff_shapes)}
\usage{
s.s.vertical(lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for a single vertical s (to be used in diff_shapes)
}
