% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eight_shapes.R
\name{s.vertical.eight}
\alias{s.vertical.eight}
\title{Define the coordinates of a single vertical eight (to be used in diff_shapes)}
\usage{
s.vertical.eight(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a single vertical eight (to be used in diff_shapes)
}
