% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.vertical.inv}
\alias{s.vertical.inv}
\title{Define the cooordinates for a vertical inverted s}
\usage{
s.vertical.inv(lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for a vertical inverted s
}
