% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.vertical}
\alias{s.vertical}
\title{Define the cooordinates for a vertical s}
\usage{
s.vertical(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{}
}
\description{
Define the cooordinates for a vertical s
}
\examples{
Arrivano
}
