% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s.vertical.s}
\alias{s.vertical.s}
\title{Coordinates of a single vertical S}
\usage{
s.vertical.s(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
Return the coordinates for drawing a vertical S
}
\description{
Define the coordinates for drawing a single vertical S composed of two arches, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# default coordinates of the vertical S
s.vertical.s()
# change the line type
s.vertical.s(lty = 2)
}
}
