% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{s_miley}
\alias{s_miley}
\title{Define the coordinates a single miley}
\usage{
s_miley(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a single miley
}
\description{
Define the coordinates of the circle sections composing a miley, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# return the default coordinates for drawing a single miley
s.miley()

# change the line type of the single miley

s.miley(lty = 3)
}
}
