% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_figures.R
\name{s_ninja}
\alias{s_ninja}
\title{Coordinates of a single ninja}
\usage{
s_ninja(size.x = 10, size.y = 15, shd = "black", lwd = 3, lty = 0)
}
\arguments{
\item{size.x}{An integer giving the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 10.}

\item{size.y}{An integer or a vector giving the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 15.}

\item{shd}{Color of the figure Default is black.}

\item{lwd}{An integer defining the width of the border line. Default is 3.}

\item{lty}{An integer defining the border line. Default is 1 (solid), can be dotted (2) or dashed (3)}
}
\value{
Return the coordinates for drawing a single ninja
}
\description{
Define the coordinates of the ellipses within which a ninja (composed of two lucks) can be inscribed, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# return the default coordinates for drawing a single ninja
s_ninja()

# change the shade of the single ninja

s_ninja(shd = "grey", lty = 0)
}
}
