% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{s_vertical_s_inv}
\alias{s_vertical_s_inv}
\title{Coordinates of a single inverted vertical S}
\usage{
s_vertical_s_inv(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
Return the coordinates for drawing a single vertical S
}
\description{
Define the coordinates for drawing a single inverted vertical S composed of two arches, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# default coordinates of the single inverted vertical S
s_vertical_s_inv()
# change the line type
s_vertical_s_inv(lty = 2)
}
}
