% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_list.R
\name{sample_resp}
\alias{sample_resp}
\alias{sample_resp.responses}
\title{Sample a random subset of responses (including correct one)}
\usage{
sample_resp(responses, n_resp = 4, ...)

\method{sample_resp}{responses}(responses, n_resp = 4, ...)
}
\arguments{
\item{responses}{A responses object (named list)}

\item{n_resp}{integer, number of responses to sample. Default is 4.}

\item{...}{other arguments}
}
\value{
A responses object of length n_resp, always including the correct
  response (index 1), with order shuffled.
}
\description{
Sample a random subset of responses (including correct one)
}
\section{Methods (by class)}{
\itemize{
\item \code{sample_resp(responses)}: Sample responses

}}
