% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections.R
\name{semi.circle.rev.inv}
\alias{semi.circle.rev.inv}
\title{Define the coordinates of an upward facing left semi-circle}
\usage{
semi.circle.rev.inv(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  theta1 = 3 * pi/4,
  theta2 = 7 * pi/4,
  shd = NA,
  rotation = pi - pi/4,
  lty = 1,
  lwd = 3,
  vis = 1
)
}
\arguments{
\item{vis}{}
}
\description{
Define the coordinates of an upward facing left semi-circle
}
