% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections_figures.R
\name{semi_circle}
\alias{semi_circle}
\title{Coordinates of a downward-facing left semi-circle}
\usage{
semi_circle(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  theta1 = pi/4,
  theta2 = 5 * pi/4,
  shd = NA,
  lty = 1,
  lwd = 3,
  vis = 1
)
}
\arguments{
\item{size.x}{The length of the x-axis. Default is sqrt(square()$ size.x[[1]]^2 /2).}

\item{size.y}{The length of the x-axis. Default is 0.}

\item{pos.x}{Position on the x axis. Default is 0.}

\item{pos.y}{Position on the y axis, Default is 0.}

\item{theta1}{Starting angle of the circle section. Default is pi/4.}

\item{theta2}{Ending angle of the circle section (built counterclockwise). Default is 5*pi/4.}

\item{shd}{Color of the figure Default is NA which results in a transparent figure.}

\item{lty}{The line type, default is 1 (solid line).}

\item{lwd}{The line width. Default is 3.}

\item{vis}{Visibility of the figure. Default is 1, making the figure visible. To hide the figure, change it to 0.}
}
\value{
The coordinates a downward-facing left semi-circle
}
\description{
Define the coordinates of a downward-facing left semi-circle
}
\examples{
\dontrun{
# default coordinates of the downward-facing left semi-circle
semi_circle()
# change the rotation of the downward-facing left semi-circle
semi_circle(theta1 = pi/2, theta2 = 3*pi/2)
}
}
