% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{shade.figure}
\alias{shade.figure}
\title{Shade figure}
\usage{
\method{shade}{figure}(fig, n = 1, rule = "shade", ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the color of the shading Default is 1 (white). Other options are 2 (grey) and 3 (black)}

\item{rule}{The rule for shading the figure}

\item{...}{Other arguments}
}
\description{
Shade figure
}
