% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{shape.figure}
\alias{shape.figure}
\title{Shape figure}
\usage{
\method{shape}{figure}(fig, n = 1, rule = "shape", ...)
}
\arguments{
\item{fig}{A vector of figures obtained with the concatenation of figures function (cof()). Three figures are needed.}

\item{n}{The number of the figure you want to see. Default is 1 (the first figure in cof() is shown). To see the other figures, change n to the position of the figure you want to show.}

\item{rule}{Define the shape rule.}

\item{...}{Other arguments}
}
\description{
Shape figure
}
