% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{shape}
\alias{shape}
\title{Change the visibility of the shapes in a figure}
\usage{
shape(fig, n, ...)
}
\arguments{
\item{fig}{A vector of figures obatined with the concatenation of figures function (`cof()`). Three figures are needed.}

\item{n}{The number of the figure you want to see. Default is 1, such that the first figure in `cof()` is shown}

\item{...}{}

\item{Define}{the non so bene come dirlo, aiuto}
}
\description{
Change the visibility of the shapes in a figure
}
