% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{show}
\alias{show}
\alias{show.figure}
\title{Show figures (Method)}
\usage{
show(obj, index)

\method{show}{figure}(obj, index = "Full")
}
\arguments{
\item{obj}{A figure composed of different figures}

\item{index}{integer, the index of the element to hide}
}
\value{
The starting object with one more visible figure

The starting object with one more visible figure
}
\description{
Change the visibility of a figure from 0 to 1
}
\section{Methods (by class)}{
\itemize{
\item \code{show(figure)}: Show figures

Change the visibility of a figure from 0 to 1

}}
\examples{
# concatenate three figures into an object. The first figure is not visible
my_shapes <- cof(square(vis = 0), triangle(), slice())
# draw object
draw(my_shapes)
# show the square
draw(show(my_shapes, 1))
# concatenate three figures into an object. The first figure is not visible
my_shapes <- cof(square(vis = 0), triangle(), slice())
# draw object
draw(my_shapes)
# show the square
draw(show(my_shapes, 1))
}
