% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{size.figure}
\alias{size.figure}
\title{Size figure}
\usage{
\method{size}{figure}(fig, n = 2, rule = "size", ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the dimension of the sizing. Default is 2.}

\item{rule}{Define the sizing rule. Default is to reduce the dimension. rule = "inv" forces to increase the dimension.}

\item{...}{Other arguments}
}
\description{
Size figure
}
