% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections.R
\name{slice}
\alias{slice}
\title{Define the coordinates of a pizza slice}
\usage{
slice(
  size.x = 15,
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  theta1 = pi/4,
  theta2 = 3 * pi/4,
  lty = 1,
  lwd = 3,
  vis = 1,
  shd = NA
)
}
\arguments{
\item{shd}{}
}
\description{
Define the coordinates of a pizza slice
}
