% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split.mat}
\alias{split.mat}
\title{Isolate ONLY the visible objects in a cell}
\usage{
\method{split}{mat}(m, cell = NULL, vis = NULL, mat.type = 9)
}
\arguments{
\item{m}{The matrix}

\item{cell}{the index of the cell to be splitted}

\item{vis}{boh}

\item{mat.type}{Does the matrix have 4 or 9 cells? Default is 9}
}
\description{
Isolate ONLY the visible objects in a cell
}
