% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_mat.figure}
\alias{split_mat.figure}
\title{Isolate the objects in a cell}
\usage{
\method{split_mat}{figure}(obj, vis = TRUE, cell = NULL)
}
\arguments{
\item{obj}{The figure}

\item{vis}{}

\item{cell}{the index of the cell to be splitted}
}
\description{
Isolate the objects in a cell
}
