% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_mat.matrix}
\alias{split_mat.matrix}
\title{Isolate the visible objects in a cell of the matrix}
\usage{
\method{split_mat}{matrix}(obj, vis = TRUE, cell = NULL)
}
\arguments{
\item{vis}{Whether you want to split only the visible figures in a cell (vis = TRUE, default) or all the figures (vis = FALSE)}

\item{cell}{The index of the cell to be splitted. Default is the cell of the correct response}

\item{m}{The matrix}
}
\value{
A list of the same length of the visible figures in the selected cell (vis = TRUE) or of the same lenght of all the figures in the matrix (vis = TRUE)
}
\description{
Isolate the visible objects in a cell of the matrix
}
\examples{
\dontrun{

my_matrix = apply(Raven(
st1 = cof(triangle(), square(), circle())
))

my_split = split_mat(my_matrix)

draw(my_split[[1]])

}

}
