% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{square.4}
\alias{square.4}
\title{Coordinates of a square composed of 4 lines}
\usage{
square.4(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = size.x,
  pos.x = size.x,
  pos.y = size.x,
  lwd = 3,
  lty = 1
)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is sqrt(square()$ size.x[[1]]^2 /2)}

\item{size.y}{numeric, define the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is size.x.}

\item{pos.x}{numeric, position on the x axis. Default is 0.}

\item{pos.y}{numeric, position the y axis, Default is 0.}

\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing an horizontal line
}
\description{
Define the coordinates for drawing a square composed of 4 lines
}
\examples{
\dontrun{
# default coordinates of square composed of 4 lines

square.4()

# draw square composed of 4 lines with different lty

draw(square.4(lty = 2))

}
}
