% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{square}
\alias{square}
\title{Define the coordinates of a square}
\usage{
square(
  size.x = 15,
  size.y = size.x,
  rot = pi/4,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\value{
Return the default square object
}
\description{
Define the coordinates of a square
}
\examples{
square()
}
