% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{square4}
\alias{square4}
\title{Define the coordinates of a square composed of 4 lines}
\usage{
square4(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = sqrt(square()$size.y[[1]]^2/2),
  pos.x = sqrt(square()$size.x[[1]]^2/2),
  pos.y = sqrt(square()$size.y[[1]]^2/2),
  lwd = 3,
  lty = 1
)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a square composed of 4 lines
}
