% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{star}
\alias{star}
\title{Define the coordinates of a star}
\usage{
star(size.x = 10, size.y = 15, shd = "black", lwd = 3, lty = 0)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a star
}
