% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{up_petal}
\alias{up_petal}
\title{Define the coordinates of an up petal}
\usage{
up_petal(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing an up petal
}
\description{
Define the coordinates of the circle sections composing an up petal
}
\examples{
\dontrun{
# return the default coordinates for drawing an up petal
up_petal()

# change the line type of the up petal

up_petal(lty = 3)
}
}
