% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_arc.R
\name{v.arc.left.down}
\alias{v.arc.left.down}
\title{Define the coordinates of the vertical left down arch}
\usage{
v.arc.left.down(
  size.x = square()$size.x[[1]]/2,
  size.y = square()$size.y[[1]]/2,
  lty = 1,
  lwd = 3,
  vis = 1,
  pos.x = 0,
  pos.y = 0
)
}
\value{
Return the vertical arc left down object
}
\description{
Define the coordinates of the vertical left down arch
}
\examples{
v.arc.left.down()
}
