% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_arc.R
\name{v.arc.left.up}
\alias{v.arc.left.up}
\title{Define the coordinates of the vertical left up arch}
\usage{
v.arc.left.up(
  size.x = square()$size.x[[1]]/2,
  size.y = square()$size.y[[1]]/2,
  pos.x = 0,
  pos.y = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{pos.y}{}
}
\description{
Define the coordinates of the vertical left up arch
}
