% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_arc.R
\name{v.arc.right.up}
\alias{v.arc.right.up}
\title{Define the coordinates of the vertical right up arch}
\usage{
v.arc.right.up(
  size.x = square()$size.x[[1]]/2,
  size.y = square()$size.y[[1]]/2,
  pos.x = 0,
  pos.y = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\value{
Return the vertical arc right up object
}
\description{
Define the coordinates of the vertical right up arch
}
\examples{
v.arc.right.up()
}
