% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eight_shapes.R
\name{vertical.eight}
\alias{vertical.eight}
\title{Define the coordinates of a vertical eight}
\usage{
vertical.eight(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a vertical eight
}
