% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eight_shapes.R
\name{vertical.s.eight}
\alias{vertical.s.eight}
\title{Coordinates of a single vertical eight}
\usage{
vertical.s.eight(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a single vertical eight to be used in shape()
}
\description{
Define the coordinates of a single vertical eight, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# default coordinates of te single vertical eight
vertical.s.eight()
# change the line type
vertical.s.eight(lty = 2)
}
}
