% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{vertical.s}
\alias{vertical.s}
\title{Coordinates of a vertical S}
\usage{
vertical.s(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
Return the coordinates for drawing a vertical S
}
\description{
Define the coordinates of a vertical S
}
\examples{
\dontrun{
# default coordinates of the vertical S
vertical.s()
# change the line type
vertical.s(lty = 2)
}
}
