% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_shapes.R
\name{vertical_s_inv}
\alias{vertical_s_inv}
\title{Coordinates of an inverted vertical S}
\usage{
vertical_s_inv(lty = 1, lwd = 3)
}
\arguments{
\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}
}
\value{
Return the coordinates for drawing an inverted vertical S
}
\description{
Define the coordinates of an inverted vertical S
}
\examples{
\dontrun{
# default coordinates of the inverted vertical S
vertical_s_inv()
# change the line type
vertical_s_inv(lty = 2)
}
}
