% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp.R
\name{wp.matriks}
\alias{wp.matriks}
\title{wp}
\usage{
\method{wp}{matriks}(obj, ...)
}
\arguments{
\item{obj}{The matriks}

\item{...}{Other arguments}
}
\value{
a list of two
}
\description{
generate wrong principle distractors
}
