% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp.R
\name{wp}
\alias{wp}
\alias{wp.matriks}
\title{Wrong principle distractors (method)}
\usage{
wp(obj, ...)

\method{wp}{matriks}(obj, ...)
}
\arguments{
\item{obj}{The matriks}

\item{...}{Other arguments}
}
\value{
An object of class responses that contains the wrong principle distractors of a matriks (WP-Matrix and WP-Copy). If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.

An object of class responses that contains the wrong principle distractors of a matriks (WP-Matrix and WP-Copy). If the distractor could not be generated because of the constraints imposed by the matrix, it will be covered by a thick, black X and a warning is given.
}
\description{
Generate the wrong principle distractors
}
\section{Methods (by class)}{
\itemize{
\item \code{wp(matriks)}: Wrong principle distractors

Generate the wrong principle distractors

}}
\examples{
m1 <- mat_apply(hexagon(),  hrules = "lty")
# draw the matriks
draw(m1)
# draw the wp distractors with the title
draw(wp(m1), main = TRUE)
m1 <- mat_apply(hexagon(),  hrules = "lty")
# draw the matriks
draw(m1)
# draw the wp distractors with the title
draw(wp(m1), main = TRUE)
}
