% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{X}
\alias{X}
\title{Define the coordinates of an X}
\usage{
X(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = sqrt(square()$size.y[[1]]^2/2),
  lwd = 3,
  lty = 1
)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of an X
}
