% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-figures.R
\name{axe}
\alias{axe}
\title{Define the cooordinates for an axe}
\usage{
axe(size.x = 15, pos.x = 0, pos.y = 0, lty = 1, lwd = 3, shd = NA)
}
\arguments{
\item{shd}{}
}
\description{
Define the cooordinates for an axe
}
