% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R, R/composed_figures.R
\name{biscuit}
\alias{biscuit}
\title{Define the coordinates of a biscuit}
\usage{
biscuit(
  pos.x = 0,
  pos.y = 0,
  size.x = 10,
  size.y = 10,
  shd = "black",
  lty = 1,
  lwd = 3
)

biscuit(
  pos.x = 0,
  pos.y = 0,
  size.x = 10,
  size.y = 10,
  shd = "black",
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{lty}{}

\item{lwd}{}
}
\value{
Based on the chosen shape, return an object with the information for plotting the desired design. If the name is precedeed by an s, the object is seen as a unique object, otherwise it is seen as a combination of multiple objects
}
\description{
Define the coordinates of a biscuit

Define the coordinates of a biscuit
}
\examples{
One day
}
