% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color.R
\name{change_color}
\alias{change_color}
\alias{change_color.figure}
\title{Change shade}
\usage{
change_color(obj, ...)

\method{change_color}{figure}(obj, ...)
}
\arguments{
\item{obj}{The figure}

\item{...}{other arguments}
}
\value{
Return the original figure with the inverted shade

Return the original figure with the inverted shade
}
\description{
Change the shade of a figure
}
\section{Methods (by class)}{
\itemize{
\item \code{change_color(figure)}: Change shade

Change the shade of a figure

}}
\examples{
# draw a square with inverted color
draw(change_color(square()))
draw(change_color(square()))
}
