% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{cof}
\alias{cof}
\title{Concatenation of figures}
\usage{
cof(..., name, single)
}
\arguments{
\item{...}{Vector of figures to be concatened together.}

\item{name}{Name of the newely created object (See Details).}

\item{single}{Should the objects be collapsed into a single object? Deafult is FALSE}
}
\value{
An object of class figure
}
\description{
Concatenation of figures
}
