% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{cross.dice}
\alias{cross.dice}
\title{Define the coordinates for drawing a cross dice with 4 dots}
\usage{
cross.dice(shd = "black", lwd = 3, lty = 1)
}
\arguments{
\item{shd}{The shading of the figure. Default is NA (transparent)}

\item{lwd}{The line width. Default is 3}

\item{lty}{The lime type, default is 1 (solid line).}
}
\description{
Define the coordinates for drawing a cross dice with 4 dots
}
