% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{cross.dice}
\alias{cross.dice}
\title{Define the coordinates of a cross dice with 4 dots}
\usage{
cross.dice(shd = "black", lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a cross dice with 4 dots
}
