% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decof.R
\name{decof}
\alias{decof}
\alias{decof.figure}
\title{Split the elements of a figure (Method)}
\usage{
decof(obj)

\method{decof}{figure}(obj)
}
\arguments{
\item{obj}{The figure of class figure to be split in its single components}
}
\value{
A named list of figures of length equal to the total of shapes in a figure (both visible and not visible)

A named list of figures of length equal to the total of shapes in a figure (both visible and not visible)
}
\description{
Return the elements composing a figure
}
\section{Methods (by class)}{
\itemize{
\item \code{decof(figure)}: Split the elements of a figure

Return the elements composing a figure

}}
\examples{
# apply the size rule on a triangle for creating a matriks with 9 cell
my_mat1 <- mat_apply(triangle(), hrules = "size")
my_mat2 <- mat_apply(dot(), hrules = "shade")
my_mat <- com(my_mat1, my_mat2)
# Return the figures composing the first cell of the matriks
decof(my_mat$Sq2)
# apply the size rule on a triangle for creating a matriks with 9 cell
my_mat1 <- mat_apply(triangle(), hrules = "size")
my_mat2 <- mat_apply(dot(), hrules = "shade")
my_mat <- com(my_mat1, my_mat2)
# Return the figures composing the first cell of the matriks
decof(my_mat$Sq2)
}
