% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{diagline.inv}
\alias{diagline.inv}
\title{Coordinates of the inverse diagonal line}
\usage{
diagline.inv(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = size.x,
  pos.x = 0,
  pos.y = 0,
  lty = 1,
  lwd = 3,
  rotation = pi + pi/4,
  vis = 1
)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is sqrt(square()$ size.x[[1]]^2 /2)}

\item{size.y}{numeric, define the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is size.x.}

\item{pos.x}{numeric, position on the x axis. Default is 0.}

\item{pos.y}{numeric, position the y axis, Default is 0.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{rotation}{integer, rotation of the line. Default is pi + pi/4.}

\item{vis}{integer, define the visibility of the figure (default is 1, visible).}
}
\value{
Return the coordinates for drawing the inverse diagonal line
}
\description{
Define the coordinates for drawing the inverse diagonal line
}
\examples{
\dontrun{
# default coordinates of the main diagonal line

diagline.inv()

# draw the main diagonal line with different lty

draw(diagline.inv(lty = 2))

}
}
