% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_figures.R
\name{dice}
\alias{dice}
\alias{cross_dice}
\title{Coordinates of a dice with four dots}
\usage{
dice(pos.x = 13, pos.y = 13, shd = "black", lwd = 3, lty = 1)

cross_dice(shd = "black", lwd = 3, lty = 1)
}
\arguments{
\item{pos.x}{numeric, position on the x axis. Default is 13 (-13)}

\item{pos.y}{numeric, position on the y axis. Default is 13 (-13)}

\item{shd}{character, define the shading of the figure. Default is black}

\item{lwd}{integer, define the line width of the figure. Default is 3}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a dice with 4 dots

The coordinates for drawing a dice with 4 dots
}
\description{
Define the coordinates for drawing four dots placed in the vertices of a square
}
\section{Functions}{
\itemize{
\item \code{cross_dice()}: Coordinates of a cross dice with four dots

Define the coordinates for drawing four dots placed in the vertices of a luck

}}
\examples{
# return the default coordinates for drawing a dot
dice()
# change the shade of the dice
dice(shd = "grey")
# return the default coordinates for drawing a dot
cross_dice()

# change the shade of the cross dice

cross_dice(shd = "grey")
}
