% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{dice}
\alias{dice}
\title{Define the coordinates for drawing a dice with 4 dots}
\usage{
dice(pos.x = 13, pos.y = 13, shd = "black", lwd = 3, lty = 1)
}
\arguments{
\item{pos.x}{The x coordinate on the Cartesian system with origin (0, 0). Deafult is 13 (-13).}

\item{pos.y}{The y coordinate on the Cartesian system with origin (0, 0). Deafult is 13 (-13).}

\item{shd}{The shading of the figure. Default is NA (transparent)}

\item{lwd}{The line width. Default is 3}

\item{lty}{The lime type, default is 1 (solid line).}
}
\description{
Define the coordinates for drawing a dice with 4 dots
}
