% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{dot}
\alias{dot}
\title{Coordinates for drawing a dot}
\usage{
dot(
  size.x = 2,
  size.y = 2,
  pos.x = 0,
  pos.y = 0,
  shd = "black",
  lty = 1,
  lwd = 3,
  vis = 1
)
}
\arguments{
\item{size.x}{The length of the x-axis. Default is 2.}

\item{size.y}{The length of the x-axis. Default is 2.}

\item{pos.x}{The x coordinate on the Cartesian system with origin (0, 0). Deafult is 0.}

\item{pos.y}{The y coordinate on the Cartesian system with origin (0, 0). Deafult is 0.}

\item{shd}{The shading of the figure. Default is black}

\item{lty}{The lime type, default is 1 (solid line).}

\item{lwd}{The line width. Default is 3}

\item{vis}{Visibility of the figure. Deafult is 1, making the figure visible. To hide the figure, change it to 0.}
}
\value{
The coordinates for drawing a black dot
}
\description{
Qui c'è una descrizione di cosa fa questo coso
}
\examples{
One day not today
}
