% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{dot}
\alias{dot}
\title{Define the coordinates of a dot}
\usage{
dot(
  size.x = 2,
  size.y = 2,
  pos.x = 0,
  pos.y = 0,
  shd = "black",
  lty = lty,
  lwd = lwd,
  vis = 1
)
}
\arguments{
\item{vis}{}
}
\description{
Define the coordinates of a dot
}
\examples{
One day not today
}
