% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{down.petal}
\alias{down.petal}
\title{Define the coordinates of a the down petal}
\usage{
down.petal(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a the down petal
}
