% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw figures}
\usage{
draw(
  obj,
  main = NULL,
  canvas = TRUE,
  hide = FALSE,
  n.figure = 9,
  bg = "white",
  mar = c(1, 1, 1, 1),
  xlim = 16
)
}
\arguments{
\item{obj}{The figure to be drawn. Can be a single figure, a matrix, or the responses}

\item{main}{logical, print the title of the drawing. Default is FALSE}

\item{canvas}{logical, draw the figure on a new canvas. Default is TRUE}

\item{hide}{logical, hide the cell corresponding to the correct response. Default is FALSE}

\item{n.figure}{integer, define the number of cells of the matrix. Default is 9}

\item{bg}{character, define the color background. Default is white}

\item{mar}{numeric vector, change margins of the canvas}

\item{xlim}{numeric, change the length of the x axis}
}
\value{
A graphic
}
\description{
Draw figures
}
\examples{
\dontrun{
# draw a circle
draw(circle())

# draw a circle inside the other
draw(size(circle(), 2), canvas = FALSE)
}
}
