% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw figures}
\usage{
draw(
  obj,
  main = NULL,
  canvas = TRUE,
  hide = FALSE,
  n.figure = 9,
  bg = "white",
  mar = c(1, 1, 1, 1),
  xlim = 16
)
}
\arguments{
\item{obj}{The object to be draw. Can be a single figure, a matrix, or the responses}

\item{main}{Print a title? Default is FALSE}

\item{canvas}{Do you want to overimpose the objects? Default is FALSE}

\item{hide}{Do you want to hide the figure of the correct response? Default is FALSE}

\item{n.figure}{How main figure should the matrix have? Default is 9}

\item{bg}{Choose the color of  the background. Deafult is white}

\item{mar}{Change margins}

\item{xlim}{Change the length of the x axis}
}
\value{
A graphic
}
\description{
Draw figures
}
