% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{ellipse}
\alias{ellipse}
\title{Define the coordinates of an ellipse}
\usage{
ellipse(
  size.x = 10,
  size.y = 7,
  rot = 0,
  shd = NA,
  pos.x = 0,
  pos.y = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\value{
Return the default ellipse object
}
\description{
Define the coordinates of an ellipse
}
\examples{
ellipse()
}
