% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{ellipse}
\alias{ellipse}
\title{Coordinates of an ellipse}
\usage{
ellipse(
  size.x = 10,
  size.y = 7,
  rot = 0,
  shd = NA,
  pos.x = 0,
  pos.y = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{An integer giving the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 10.}

\item{size.y}{An integer or a vector giving the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 7.}

\item{rot}{Rotation of the figure. Default is 0.}

\item{shd}{Color of the figure Default is NA which results in a transparent figure}

\item{pos.x}{Position on the x axis. Default is 0.}

\item{pos.y}{Position on the y axis, Default is 0.}

\item{vis}{Integer, indicates whether the object should be visible (1) or not (0). Default is 1}

\item{lty}{An integer defining the border line. Default is 1 (solid), can be dotted (2) or dashed (3)}

\item{lwd}{An integer defining the width of the border line. Default is 3.}
}
\value{
Return the coordinates fro drawing a ellipse
}
\description{
Define the coordinates an ellipse
}
\examples{
\dontrun{
# return the default coordinates for drawing an ellipse
ellipse()

# change the coordinates for drawing a smaller ellipse

ellipse(size.x = 5, size.y = 3)
}
}
