% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{hexagon}
\alias{hexagon}
\title{Coordinates of a hexagon}
\usage{
hexagon(
  size.x = 15,
  size.y = size.x,
  rot = 0,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{numeric, define the semi-minor axis of the ellipse within which the figure is inscribed. Default is 15}

\item{size.y}{numeric, define the semi-minor axis of the ellipse within which the figure is inscribed. Default is size.x}

\item{rot}{define the rotation. Default is 0}

\item{pos.x}{numeric, position on the x axis. Default is 0}

\item{pos.y}{numeric, position the y axis, Default is 0}

\item{shd}{character, define the shading of the figure. Default is NA which results in a transparent figure}

\item{vis}{Visibility of the figure. Default is 1, making the figure visible. To hide the figure, change it to 0}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3}
}
\value{
Return the coordinates for drawing an hexagon
}
\description{
Define the coordinates for drawing an hexagon
}
\examples{
# return the default coordinates for drawing a hexagon
hexagon()
# change the coordinates for drawing a smaller hexagon
hexagon(size.x = 10)
}
