% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{hide}
\alias{hide}
\title{Hide figures (Method)}
\usage{
hide(obj, index)
}
\arguments{
\item{obj}{A figure composed of different figures}

\item{index}{integer, the index of the element to hide}
}
\value{
The starting object with a hidden figure
}
\description{
Change the visibility of a figure from 1 to 0
}
\examples{
# concatenate three figures into an object
my_shapes <- cof(square(), triangle(), slice())
# draw object
draw(my_shapes)
# hide the triangle
draw(hide(my_shapes, 2))
}
