% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eight_shapes.R
\name{horizontal.eight}
\alias{horizontal.eight}
\title{Define the coordinates of an horizontal eight}
\usage{
horizontal.eight(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of an horizontal eight
}
