% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{identity}
\alias{identity}
\title{Identity rule}
\usage{
identity(fig, ...)
}
\arguments{
\item{fig}{The figures on which the rules is applied.}

\item{...}{}
}
\value{
An object composed of figures combined according to an identity rule
}
\description{
Apply an identity rule (i.e., no changes)
}
