% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{left.petal}
\alias{left.petal}
\title{Define the coordinates of a left petal}
\usage{
left.petal(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a left petal
}
\description{
Define the coordinates of the circle sections composing a left petal
}
\examples{
\dontrun{
# return the default coordinates for drawing a left petal
left.petal()

# change the line type of the left petal

left.petal(lty = 3)
}
}
