% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{lily}
\alias{lily}
\alias{s_lily}
\title{Define the coordinates of a lily}
\usage{
lily(lwd = 3, lty = 1)

s_lily(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line)}
}
\value{
Return the coordinates for drawing the circle arches composing a lily

Return the coordinates for drawing the circle arches composing a single lily, to be used in shape()
}
\description{
Define the coordinates for drawing the circle arches composing a lily
}
\section{Functions}{
\itemize{
\item \code{s_lily()}: Define the coordinates a single lily

Define the coordinates for drawing the circle arches composing a single lily, to be used in shape()

}}
\examples{
# return the default coordinates drawing the circle arches composing a lily
lily()
# change the line type of the lily
lily(lty = 3)
# return the default coordinates for drawing a single lily
s_lily()
# change the line type of the single lily
s_lily(lty = 3)
}
