% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{logic}
\alias{logic}
\title{Apply logic rules to different figures}
\usage{
logic(fig, n, rule, seed, ...)
}
\arguments{
\item{fig}{A vector of figures obatined with the concatenation of figures function (`cof()`). Three figures are needed.}

\item{n}{???}

\item{rule}{Define the logic rule to be applied, either `AND`, `OR`, `XOR`}

\item{seed}{Set the ranodm seed so that the permutations are consistent}

\item{...}{}
}
\description{
Apply logic rules to different figures
}
