% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{luck.4}
\alias{luck.4}
\title{Define the coordinates of a luck composed of 4 lines}
\usage{
luck.4(size.x = 10, size.y = 7, lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a luck composed of 4 lines
}
