% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{luck}
\alias{luck}
\title{Define the coordinates of a luck}
\usage{
luck(
  size.x = 10,
  size.y = 15,
  rot = pi/2,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{lwd}{}
}
\description{
Define the coordinates of a luck
}
