% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-figures.R
\name{maxi}
\alias{maxi}
\title{Define the cooordinates for a maxi}
\usage{
maxi(size.x = 8, size.y = 4, pos.x = 0, shd = NA, lty = 1, lwd = 3)
}
\arguments{
\item{lwd}{}
}
\description{
Define the cooordinates for a maxi
}
