% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{miley}
\alias{miley}
\title{Define the coordinates of a miley}
\usage{
miley(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a miley made of petals
}
\description{
Define the coordinates of the circle sections composing a miley
}
\examples{
\dontrun{
# return the default coordinates for drawing a right petal
miley()

# change the line type of the right petal

miley(lty = 3)
}
}
