% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_sections.R
\name{pacman}
\alias{pacman}
\title{Define the coordinates of a pacman}
\usage{
pacman(
  size.x = sqrt(square()$size.x[[1]]^2/2),
  size.y = 0,
  pos.x = 0,
  pos.y = 0,
  theta1 = pi/4,
  theta2 = 7 * pi/4,
  lty = 1,
  lwd = 3,
  shd = NA,
  vis = 1
)
}
\arguments{
\item{vis}{}
}
\description{
Define the coordinates of a pacman
}
