% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_figures.R
\name{phantom}
\alias{phantom}
\title{Coordinates of a panthom figure}
\usage{
phantom()
}
\value{
An empty figure (nothing is plotted in draw)
}
\description{
Draw an empty figure
}
\examples{
# empty figure
phantom()
# draw an empty figure
draw(phantom())
}
