% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{rectangle}
\alias{rectangle}
\title{Define the coordinates of a rectangle}
\usage{
rectangle(
  size.x = 15,
  size.y = 20,
  rot = pi/4,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  lwd = 3,
  lty = 1,
  vis = 1
)
}
\arguments{
\item{vis}{}
}
\description{
Define the coordinates of a rectangle
}
