% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{reflect}
\alias{reflect}
\title{Reflect a figure}
\usage{
reflect(fig, n, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the reflection. Default is 2 which corresponds to a 180 degree reflection (CREDO)}

\item{...}{}
}
\description{
Reflect a figure
}
