% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{reflect}
\alias{reflect}
\title{Reflection rule}
\usage{
reflect(fig, n, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the reflection. Default is 2 which corresponds to a 180 degree (pi) reflection}

\item{...}{}
}
\value{
A figure with different rotation coordinates
}
\description{
Apply a reflection of pi on a figure
}
\examples{
\dontrun{
# default pacman
draw(pacman())

# apply the default rotation on the default luck
draw(reflect(pacman()))
}
}
