% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_figure.R
\name{replace}
\alias{replace}
\title{Replace objects}
\usage{
replace(obj, index, replacement)
}
\arguments{
\item{obj}{The figure of a matrix}

\item{index}{The index of the element to replace}

\item{replacement}{}
}
\value{
An object with a changed figure
}
\description{
Replace an object in a figure with another object
}
