% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_list.R
\name{response_list}
\alias{response_list}
\title{Response list (method)}
\usage{
response_list(obj, seed, ...)
}
\arguments{
\item{obj}{the matrix}

\item{seed}{the seed}

\item{...}{other arguments}
}
\value{
A list
}
\description{
Response list (method)
}
