% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{right.petal}
\alias{right.petal}
\title{Define the coordinates for drawing the right petal}
\usage{
right.petal(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{The line width. Default is 3}

\item{lty}{The lime type, default is 1 (solid line).}
}
\description{
Define the coordinates for drawing the right petal
}
