% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{right.petal}
\alias{right.petal}
\title{Define the coordinates of a the right petal}
\usage{
right.petal(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a the right petal
}
