% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rotate}
\alias{rotate}
\title{Rotation rule}
\usage{
rotate(fig, n, rule, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the angle of the rotation. Default is 4, which corresponds to a rotation of pi/4.}

\item{rule}{Define the rotation rule. Default is counterclockwise.`rule = "inv"` forces a clockwise rotation}

\item{...}{Other arguments}
}
\value{
A figure with different rotation coordinates
}
\description{
Apply a rotation of pi/4 to a figure.
}
\examples{
\dontrun{
# default luck
draw(luck())

# apply the default rotation on the default luck
draw(rotate(luck()))

# force clockwise rotation
draw(rotate(luck(), rule = "inv"))
}
}
