% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_figures.R
\name{s.axe}
\alias{s.axe}
\title{Coordinates of a single axe}
\usage{
s.axe(size.x = 15, pos.x = 0, pos.y = 0, lty = 1, lwd = 3, shd = NA)
}
\arguments{
\item{size.x}{numeric, define the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 10.}

\item{pos.x}{numeric, define the position on the x axis. Default is 0.}

\item{pos.y}{numeric, position the y axis, Default is 0.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}

\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{shd}{character, define the shading of the figure. Default is NA which results in a transparent figure}

\item{vis}{integer, define the visibility of the figure (default is 1, visible).}
}
\value{
Return the coordinates for drawing an axe
}
\description{
Define the coordinates for drawing a single axe, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# return the default coordinates for drawing an axe
s.axe()

# change the coordinates for drawing a smaller single Malta cross

s.axe(size.x = 5)
}
}
