% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{s.flower}
\alias{s.flower}
\title{Define the coordinates of a single flower (to be used in diff shapes)}
\usage{
s.flower(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a single flower (to be used in diff shapes)
}
