% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{s.flower}
\alias{s.flower}
\title{Define the coordinates for drawing a single flower (to be used in `shape()`)}
\usage{
s.flower(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{The line width. Default is 3}

\item{lty}{The lime type, default is 1 (solid line).}
}
\description{
Define the coordinates for drawing a single flower (to be used in `shape()`)
}
