% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{s.lily}
\alias{s.lily}
\title{Define the coordinates a single lily}
\usage{
s.lily(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3.}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line).}
}
\value{
Return the coordinates for drawing a single lily
}
\description{
Define the coordinates of the circle sections composing a lily, which is forced to be a single figure (to be used in shape())
}
\examples{
\dontrun{
# return the default coordinates for drawing a single lily
s.lily()

# change the line type of the single lily

s.lily(lty = 3)
}
}
