% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers.R
\name{s.lily}
\alias{s.lily}
\title{Define the coordinates of a a single lily (to be used in diff_shapes)}
\usage{
s.lily(lwd = 3, lty = 1)
}
\arguments{
\item{lty}{}
}
\description{
Define the coordinates of a a single lily (to be used in diff_shapes)
}
