% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{shade}
\alias{shade}
\title{Shading of a figure}
\usage{
shade(fig, n, rule, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the color of the shading Default is 1 (white). Other options are 2 (grey) and 3 (black)}

\item{rule}{The rule for shading the figure}

\item{...}{Other arguments}
}
\value{
A figure with different shading characteristics
}
\description{
Change the shading of a figure
}
\examples{
\dontrun{

# draw default triangle
draw(triangle())

# make it grey
draw(shade(triangle(), 2))

}
}
