% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{shape}
\alias{shape}
\title{Change the visibility of the shapes in a figure}
\usage{
shape(fig, n, ...)
}
\arguments{
\item{fig}{A vector of figures obtained with the concatenation of figures function (cof()). Three figures are needed.}

\item{n}{The number of the figure you want to see. Default is 1 (the first figure in cof() is shown). To see the other figures, change n to the position of the figure you want to show.}

\item{...}{}

\item{Define}{the non so bene come dirlo, aiuto}
}
\value{
A list of three figures, only the first of which is visible
}
\description{
Change the visibility of the shapes in a figure
}
\examples{
\dontrun{
# Three figures, only the first is shown
draw(shape(cof(s.lily(), square(), s.star())))

# Show the third figure (star)
draw(shape(cof(s.lily(), square(), s.star()), n = 3))

# Show the first and the second figures
 draw(shape(cof(s.lily(), square(), s.star()), n = c(1,2)))
}
}
