% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_figure.R
\name{show}
\alias{show}
\title{Show figures}
\usage{
show(obj, index)
}
\arguments{
\item{obj}{The figure of a matrix}

\item{index}{The index of the element to show}
}
\value{
An object with changed visibility
}
\description{
Change the visibility of the single objects composing a figure from 0 to 1
}
