% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{size}
\alias{size}
\title{Resize a figure}
\usage{
size(fig, n, rule, ...)
}
\arguments{
\item{fig}{The figure on which the rule is applied}

\item{n}{A number defining the dimension of the sizing. Default is 2, which corresponds to a change of 1.8 COSA?? BANANE???}

\item{rule}{Define the sizing rule. Default is to reduce the dimension. `rule = "inv"` forces to increase the dimension.}

\item{...}{}
}
\description{
Resize a figure
}
