% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{square}
\alias{square}
\title{Coordinates of a square}
\usage{
square(
  size.x = 15,
  size.y = size.x,
  rot = pi/4,
  pos.x = 0,
  pos.y = 0,
  shd = NA,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{An integer giving the semi-major axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is 15.}

\item{size.y}{An integer or a vector giving the semi-minor axis of the ellipse for the polygon(s) within which the figure is inscribed. Default is size.x.}

\item{rot}{Rotation of the figure. Default is pi/4.}

\item{pos.x}{Position on the x axis. Default is 0.}

\item{pos.y}{Position on the y axis, Default is 0.}

\item{shd}{Color of the figure Default is NA which results in a transparent figure}

\item{vis}{Integer, indicates whether the object should be visible (1) or not (0). Default is 1}

\item{lty}{An integer defining the border line. Default is 1 (solid), can be dotted (2) or dashed (3)}

\item{lwd}{An integer defining the width of the border line. Default is 3.}
}
\value{
Return the coordinates for drawing a square
}
\description{
Define the coordinates  of the ellipse within which a square can be inscribed.
}
\examples{
\dontrun{
# return the default coordinates for drawing a square
square()

# change the coordinates for drawing a smaller square

square(size.x = 5)
}
}
