% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black-figures.R
\name{star}
\alias{star}
\title{Define the coordinates for drawing a star}
\usage{
star(size.x = 10, size.y = 15, shd = "black", lwd = 3, lty = 0)
}
\arguments{
\item{size.x}{The length of the x-axis. Default is 10.}

\item{size.y}{The length of the x-axis. Default is 15.}

\item{shd}{The shading of the figure. Default is black}

\item{lwd}{The line width. Default is 3}

\item{lty}{The lime type, default is 1 (solid line).}
}
\description{
Define the coordinates for drawing a star
}
