% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_figures.R
\name{triangle}
\alias{triangle}
\title{Define the coordinates of a triangle}
\usage{
triangle(
  size.x = 15,
  size.y = 15,
  pos.x = 0,
  pos.y = 0,
  rot = pi/2,
  shd = NA,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\value{
Return the default triangle object
}
\description{
Define the coordinates of a triangle
}
\examples{
triangle()
}
