% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowers_figures.R
\name{up_petal}
\alias{up_petal}
\alias{down_petal}
\alias{left_petal}
\alias{right_petal}
\title{Define the coordinates of petals}
\usage{
up_petal(lwd = 3, lty = 1)

down_petal(lwd = 3, lty = 1)

left_petal(lwd = 3, lty = 1)

right_petal(lwd = 3, lty = 1)
}
\arguments{
\item{lwd}{integer, define the line width of the figure. Default is 3}

\item{lty}{integer, define the line type of the figure, default is 1 (solid line)}
}
\value{
Return the coordinates for drawing the circle arches composing an up petal

Return the coordinates for drawing  the circle arches composing a down petal

Return the coordinates for drawing the circle arches composing a left petal

Return the coordinates for drawing the circle arches composing a right petal
}
\description{
Define the coordinates for drawing the circle arches composing some petals
}
\section{Functions}{
\itemize{
\item \code{down_petal()}: Define the coordinates of a down petal

Define the coordinates for drawing the circle arches composing a down petal

\item \code{left_petal()}: Define the coordinates of a left petal

Define the coordinates for drawing the circle arches composing a left petal

\item \code{right_petal()}: Define the coordinates of a right petal

Define the coordinates for drawing the circle arches composing a right petal

}}
\examples{
# return the default coordinates for drawing the circle arches composing an up petal
up_petal()
# change the line type of the up petal
up_petal(lty = 3)
# return the default coordinates for drawing a down petal
down_petal()
# change the line type of the down petal
down_petal(lty = 3)
# return the default coordinates for drawing a left petal
left_petal()
# change the line type of the left petal
left_petal(lty = 3)
# return the default coordinates for drawing a right petal
right_petal()
# change the line type of the right petal
right_petal(lty = 3)
}
