% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_arc.R
\name{v.arc.right.up}
\alias{v.arc.right.up}
\title{Define the coordinates for drawing the vertical right up arch}
\usage{
v.arc.right.up(
  size.x = square()$size.x[[1]]/2,
  size.y = square()$size.y[[1]]/2,
  pos.x = 0,
  pos.y = 0,
  vis = 1,
  lty = 1,
  lwd = 3
)
}
\arguments{
\item{size.x}{The length of the x-axis. Default is square()$size.x[[1]]/2.}

\item{size.y}{The length of the x-axis. Default is square()$size.y[[1]]/2.}

\item{pos.x}{The x coordinate on the Cartesian system with origin (0, 0). Deafult is 0.}

\item{pos.y}{The y coordinate on the Cartesian system with origin (0, 0). Deafult is 0.}

\item{vis}{Visibility of the figure. Deafult is 1, making the figure visible. To hide the figure, change it to 0.}

\item{lty}{The lime type, default is 1 (solid line).}

\item{lwd}{The line width. Default is 3}
}
\value{
Return the vertical arc right up object
}
\description{
Define the coordinates for drawing the vertical right up arch
}
\examples{
v.arc.right.up()
}
